fix: ESP-IDF 6.x TWAI API migration, disable WiFi, throttle cast
- Replace deprecated TWAI_TIMING_CONFIG_500KBITS macro with literal values
- Use TWAI_GENERAL_CONFIG_DEFAULT_V2 for ESP-IDF 6.x compatibility
- Add esp_timer.h include to RT main.cpp for esp_timer_get_time
- Fix SysThrottleSts int32->int16 narrowing cast
- Add sdkconfig.defaults to disable WiFi (wpa_supplicant errors)
- Clean .pio build directories for fresh config

+++ sys-esp32/src/main.cpp
@@ -218,7 +218,7 @@ static QueueHandle_t g_setpoint_queue = nullptr;  // 4 deep, ActuatorSetpoint
 
         // Send 0x120 SYS_THROTTLE_STS
         can::Frame fr;
-        can::SysThrottleSts{g_throttle.read_mmps()}.to_frame(fr);
+        can::SysThrottleSts{int16_t(g_throttle.read_mmps())}.to_frame(fr);
         g_can.send(fr);
 
         vTaskDelayUntil(&last, period);
